requirements.txt
Appium-Python-Client
selenium
pydub
fastapi
uvicorn
python-multipart
